Only attempt to remove databases if they exist

+++ Publish/CI_Cleanup.ps1
@@ -125,23 +125,67 @@ else
 #endregion
 
 #region Drop Databases
-log_statement "Deleting application database"
-$command = "'c:\program` files\postgresql\9.6\bin\psql.exe' -d postgres -h $dbServer -e -q --command=`"drop database $appDbName`""
-invoke-expression "&$command"
+
+$env:PGPASSWORD = $dbPassword
+
+# Check if databases already exist
+$appDbFound = $false
+$logDbFound = $false
+
+$command = "$psqlExePath --dbname=postgres  -h $dbServer -U $dbUser --tuples-only --command=`"select datname from Pg_database`" --echo-errors"
+$output = invoke-expression "&$command"
 
 if ($LASTEXITCODE -ne 0) {
     $error_code = $LASTEXITCODE
-    log_statement "ERROR: Failed to drop application database"
-    log_statement $requestResult.stdout
+    log_statement "ERROR: Failed to query for existing databases"
+    log_statement "errorlevel was $LASTEXITCODE"
+    exit $error_code
 }
 
-log_statement "Deleting logging database"
-$command = "'c:\program` files\postgresql\9.6\bin\psql.exe' -d postgres -h $dbServer -e -q --command=`"drop database $logDbName`""
-invoke-expression "&$command"
+foreach ($db in $output) {
+    if ($db.trim() -eq $appDbName) {
+        log_statement "MAP application database found for this branch."
+        $appDbFound = $true
+    }
+    elseif ($db.trim() -eq $logDbName) {
+        log_statement "Logging database found for this branch."
+        $logDbFound = $true
+    }
+}
 
-if ($LASTEXITCODE -ne 0) {
-    $error_code = $LASTEXITCODE
-    log_statement "ERROR: Failed to drop logging database"
-    log_statement $requestResult.stdout
+if ($appDbFound)
+{
+    log_statement "Deleting application database"
+    $command = "'c:\program` files\postgresql\9.6\bin\psql.exe' -d postgres -h $dbServer -e -q --command=`"drop database $appDbName`""
+    invoke-expression "&$command"
+
+    if ($LASTEXITCODE -ne 0) {
+        $error_code = $LASTEXITCODE
+        log_statement "ERROR: Failed to drop application database"
+        log_statement $requestResult.stdout
+        exit 42
+    }
+}
+else 
+{
+    log_statement "Application database was not found for this branch"
+}
+
+if ($logDbFound)
+{
+    log_statement "Deleting logging database"
+    $command = "'c:\program` files\postgresql\9.6\bin\psql.exe' -d postgres -h $dbServer -e -q --command=`"drop database $logDbName`""
+    invoke-expression "&$command"
+
+    if ($LASTEXITCODE -ne 0) {
+        $error_code = $LASTEXITCODE
+        log_statement "ERROR: Failed to drop audit log database"
+        log_statement $requestResult.stdout
+        exit 42
+    }
+}
+else
+{
+    log_statement "Audit Log database was not found for this branch"
 }
 #endregion
